Withdraws and Deposits
Added code to withdraw and deposit money at the request of a command. Added new table column containing the amount withdrew and one for the amount deposited today. This allows it to require approval if it goes above that number. All I need to do now is code it to actually reset that number.