Data Edit Command and Permissions
Added a system where only people whose user ID is in a list called ADMINS in the .env file can run certain commands, such as the new accountEdit command. This command lets you edit and piece of data in any account. I could further generalize it to any table in the data, but since there is only one table currently, there is not much of a point.