Lottery table, ticket buying, and lottery ending
There is a table called lottery than only stores account names. When you use the buy lottery ticket command, it buys one and sets its account name to the one used as a parameter in the command. It transfers a percentage of that to the IMC account and the rest of it to the Lottery account. Then, an admin can run the end lottery command to end the lottery, pick a random winner, transfer all the money in the lottery account to the winner's accounts, and clear the lottery database.